# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mina REST API
  version: 0.1.0
paths:
...
  /tags:
    get:
      operationId: listTags
      summary: List tags.
      parameters:
        - name: include_hidden
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: fqn
            enum:
              - fqn
              - created_at
              - updated_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Tags in deterministic FQN order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    post:
      operationId: createTag
      summary: Create a tag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /tags/{tag_id}:
    parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getTag
      summary: Get a tag.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Tag found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTag
      summary: Update tag hidden state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTagRequest"
      responses:
        "200":
          description: Tag updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTag
      summary: Tombstone a tag.
      responses:
        "204":
          description: Tag tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /members:
    get:
      operationId: listMembers
      summary: List household members.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name
            enum:
              - name
              - created_at
              - updated_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Members in deterministic name order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    post:
      operationId: createMember
      summary: Create a household member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
      responses:
        "201":
          description: Member created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /members/{member_id}:
    parameters:
      - name: member_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getMember
      summary: Get a household member.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Member found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateMember
      summary: Update household member name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Member updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteMember
      summary: Tombstone a household member.
      responses:
        "204":
          description: Member tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /exchange-rates:
    get:
      operationId: listExchangeRates
      summary: List exchange rates.
      parameters:
        - name: from_currency
          in: query
          required: false
          schema:
            type: string
            pattern: ^[A-Z]{3}$
        - name: to_currency
          in: query
          required: false
          schema:
            type: string
            pattern: ^[A-Z]{3}$
        - name: effective_date
          in: query
          required: false
          schema:
            type: string
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: currency_pair
            enum:
              - currency_pair
              - from_currency
              - to_currency
              - effective_date
              - created_at
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Exchange rates in deterministic currency/date order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    post:
      operationId: createExchangeRate
      summary: Create an exchange rate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExchangeRateRequest"
      responses:
        "201":
          description: Exchange rate created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /exchange-rates/{exchange_rate_id}:
    parameters:
      - name: exchange_rate_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getExchangeRate
      summary: Get an exchange rate.
      parameters:
        - name: include_tombstoned
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Exchange rate found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateExchangeRate
      summary: Update an exchange rate value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExchangeRateRequest"
      responses:
        "200":
          description: Exchange rate updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteExchangeRate
      summary: Tombstone an exchange rate.
      responses:
        "204":
          description: Exchange rate tombstoned.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    InvalidRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ExchangeRate:
      type: object
      additionalProperties: false
      required:
        - exchange_rate_id
        - from_currency
        - to_currency
        - rate
        - effective_date
        - created_at
      properties:
        exchange_rate_id:
          type: integer
          format: int64
        from_currency:
          type: string
          pattern: ^[A-Z]{3}$
        to_currency:
          type: string
          pattern: ^[A-Z]{3}$
        rate:
          type: string
          pattern: ^[0-9]{1,18}(\.[0-9]{1,8})?$
          maxLength: 19
        effective_date:
          type: string
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
        created_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    ExchangeRateListResponse:
      type: object
      additionalProperties: false
      required:
        - exchange_rates
      properties:
        exchange_rates:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeRate"
    APIError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - not_found
            - method_not_allowed
            - conflict
            - internal_error
        message:
          type: string
...
    CreateExchangeRateRequest:
      type: object
      additionalProperties: false
      required:
        - from_currency
        - to_currency
        - rate
        - effective_date
      properties:
        from_currency:
          type: string
          pattern: ^[A-Z]{3}$
        to_currency:
          type: string
          pattern: ^[A-Z]{3}$
        rate:
          type: string
          pattern: ^[0-9]{1,18}(\.[0-9]{1,8})?$
          maxLength: 19
        effective_date:
          type: string
          pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
    CreateTagRequest:
      type: object
      additionalProperties: false
      required:
        - fqn
      properties:
        fqn:
          type: string
        is_hidden:
          type: boolean
          default: false
    CreateMemberRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/APIError"
...
    Member:
      type: object
      additionalProperties: false
      required:
        - member_id
        - name
        - created_at
        - updated_at
      properties:
        member_id:
          type: integer
          format: int64
        name:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    MemberListResponse:
      type: object
      additionalProperties: false
      required:
        - members
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    Tag:
      type: object
      additionalProperties: false
      required:
        - tag_id
        - fqn
        - is_hidden
        - name
        - level
        - created_at
        - updated_at
      properties:
        tag_id:
          type: integer
          format: int64
        fqn:
          type: string
        is_hidden:
          type: boolean
        parent_fqn:
          type: string
          nullable: true
        name:
          type: string
        level:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        tombstoned_at:
          type: string
          nullable: true
    TagListResponse:
      type: object
      additionalProperties: false
      required:
        - tags
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
...
    UpdateExchangeRateRequest:
      type: object
      additionalProperties: false
      required:
        - rate
      properties:
        rate:
          type: string
          pattern: ^[0-9]{1,18}(\.[0-9]{1,8})?$
          maxLength: 19
    UpdateTagRequest:
      type: object
      additionalProperties: false
      required:
        - is_hidden
      properties:
        is_hidden:
          type: boolean
    UpdateMemberRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string